hippocampus/requirements.txt
openai
letta
python-dateutil 